requirements.txt
sentence-transformers>=2.2.2
faiss-gpu==1.7.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
nltk>=3.8.1
scikit-learn>=1.3.0
tqdm>=4.65.0 